SRD: dont create a new sendEncodingParameters
do not create sendEncodingParameters on SRD. This could go wrong in the following scenario:
    createOffer-SLD-SRD-SRD(type=offer)
and leads to InvalidAccessErrors as the SSRC changes. Also ensures createOffer and createAnswer
work well in this scenario and reuse rather than recreating.

Also changes the RTX ssrc generation be "ssrc for this mline + 1" for consistency.